Updated documentation for building instructions
Add the required swiftlint as dependency and explain how to build
the project. Also update the project for M1 as homebrew path is
not read leading to build errors. If we detect M1, we add the
current homebrew path (/opt/homebrew).

Unfortunately, this is a dirty hack and should be fixed properly
with minimal shell scripting in near future.